+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## (unreleased)
 
+- Added a date field to the calendar.
+  - The date field will try it's best to parse the value, if no value can be parsed then it will set to the current date 
+- Fixed errors with budget state.
+- Changed date formatting to use British locale.
 - Added a new collapsible sidebar with improved UI and UX. 
   - The new sidebar has a list of all the accounts, as well as a button for creating accounts.